Extract common viem setup code to separte file
Note that in many cases we were able to remove explicit use of account
in transaction creattion because it is "hoisted" in the walletClient.

Also removed chain-id and ACC_PRIVKEY as arguments from send_tx.mjs
because they are controlled by env vars, which are picked up by the mjs
file.
