+++ e2e_test/js-tests/test_viem_tx.mjs
@@ -1,48 +1,9 @@
 import { assert } from "chai";
 import "mocha";
 import {
-	createPublicClient,
-	createWalletClient,
-	http,
-	defineChain,
 	parseAbi,
 } from "viem";
-import { celoAlfajores } from "viem/chains";
-import { privateKeyToAccount } from "viem/accounts";
-
-// Setup up chain
-const devChain = defineChain({
-	...celoAlfajores,
-	id: 1337,
-	name: "local dev chain",
-	network: "dev",
-	rpcUrls: {
-		default: {
-			http: [process.env.ETH_RPC_URL],
-		},
-	},
-});
-
-const chain = (() => {
-	switch (process.env.NETWORK) {
-		case 'alfajores':
-			return celoAlfajores
-		default:
-			return devChain
-	};
-})();
-
-// Set up clients/wallet
-const publicClient = createPublicClient({
-	chain: chain,
-	transport: http(),
-});
-const account = privateKeyToAccount(process.env.ACC_PRIVKEY);
-const walletClient = createWalletClient({
-	account,
-	chain: chain,
-	transport: http(),
-});
+import {publicClient, walletClient} from "./viem_setup.mjs"
 
 // Returns the base fee per gas for the current block multiplied by 2 to account for any increase in the subsequent block.
 async function getGasFees(publicClient, tip, feeCurrency) {
@@ -60,13 +21,12 @@ const testNonceBump = async (
 	shouldReplace,
 ) => {
 	const syncBarrierRequest = await walletClient.prepareTransactionRequest({
-		account,
+
 		to: "0x00000000000000000000000000000000DeaDBeef",
 		value: 2,
 		gas: 22000,
 	});
 	const firstTxHash = await walletClient.sendTransaction({
-		account,
 		to: "0x00000000000000000000000000000000DeaDBeef",
 		value: 2,
 		gas: 171000,
@@ -78,7 +38,6 @@ const testNonceBump = async (
 	var secondTxHash;
 	try {
 		secondTxHash = await walletClient.sendTransaction({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 3,
 			gas: 171000,
@@ -115,7 +74,6 @@ const testNonceBump = async (
 describe("viem send tx", () => {
 	it("send basic tx and check receipt", async () => {
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 1,
 			gas: 21000,
@@ -130,7 +88,6 @@ describe("viem send tx", () => {
 
 	it("send basic tx using viem gas estimation and check receipt", async () => {
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 1,
 		});
@@ -145,7 +102,6 @@ describe("viem send tx", () => {
 	it("send fee currency tx with explicit gas fields and check receipt", async () => {
 		const [maxFeePerGas, tip] = await getGasFees(publicClient, 2n, process.env.FEE_CURRENCY);
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 2,
 			gas: 171000,
@@ -163,7 +119,6 @@ describe("viem send tx", () => {
 
 	it("test gas price difference for fee currency", async () => {
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 2,
 			gas: 171000,
@@ -220,7 +175,6 @@ describe("viem send tx", () => {
 
 	it("send fee currency with gas estimation tx and check receipt", async () => {
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 2,
 			feeCurrency: process.env.FEE_CURRENCY,
@@ -286,7 +240,6 @@ describe("viem send tx", () => {
 
 	it("send tx with unregistered fee currency", async () => {
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 2,
 			gas: 171000,
@@ -334,7 +287,6 @@ describe("viem send tx", () => {
 			assert.fail(`Converted base fee (${convertedBaseFee}) not less than native base fee (${block.baseFeePerGas})`);
 		}
 		const request = await walletClient.prepareTransactionRequest({
-			account,
 			to: "0x00000000000000000000000000000000DeaDBeef",
 			value: 2,
 			gas: 171000,
